fix(release): sanitize crate manifests for publish
