+++ tools/moon/cmd/publish_crates/moon.pkg
@@ -6,6 +6,7 @@ import {
   "moonbitlang/async",
   "moonbitlang/async/process",
   "moonbitlang/async/stdio",
+  "moonbitlang/x/fs" @xfs,
   "moonbitlang/x/sys",
 }
